// web/src/lib/api/model/jobStageStatus.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Silo
 * OpenAPI spec version: 0.1.0
 */

export type JobStageStatus = (typeof JobStageStatus)[keyof typeof JobStageStatus]

export const JobStageStatus = {
  pending: "pending",
  running: "running",
  completed: "completed",
  failed: "failed",
} as const